refactor: remove redundant code of search helper.
move search helper from /server/helpers to /common/helpers and update the accordingly.
