+++ src/server/routes/search.js
@@ -21,7 +21,7 @@ import * as auth from '../helpers/auth';
 import * as error from '../../common/helpers/error';
 import * as handler from '../helpers/handler';
 import * as propHelpers from '../../client/helpers/props';
-import * as search from '../helpers/search';
+import * as search from '../../common/helpers/search';
 import * as utils from '../helpers/utils';
 
 import {keys as _keys, snakeCase as _snakeCase, isNil} from 'lodash';
